# assets/swagger/swagger.v1.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This documentation describes the GoT API.",
    "title": "GoT API.",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "paths": {
    "/v1/ping": {
      "get": {
        "produces": [
          "text/plain"
        ],
        "summary": "Echo pong to a ping",
        "operationId": "ping",
        "responses": {
          "200": {
            "$ref": "#/responses/PongResponse"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/v1/tilejson": {
      "get": {
        "produces": [
          "text/plain"
        ],
        "summary": "Return tilejson definition",
        "operationId": "tilejson",
        "responses": {
          "200": {
            "$ref": "#/responses/TileJSON"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/v1/{z}/{x}/{y}.{ext}": {
      "get": {
        "produces": [
          "image/png",
          "application/vnd.mapbox-vector-tile",
          "application/x-protobuf"
        ],
        "summary": "Map tile",
        "operationId": "getTile",
        "parameters": [
          {
            "type": "integer",
            "description": "the zoom",
            "name": "z",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "the x coord",
            "name": "x",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "the y coord",
            "name": "y",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MapTile"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "responses": {
    "MapTile": {
      "description": "MapTile represent a png map tile response"
    },
    "PongResponse": {
      "description": "PongResponse represent a pong response"
    }
  }
}